// src/LoDb.Web/src/app/core/api/generated/models/rune-row.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RuneEntry } from '../models/rune-entry';

/**
 * A row of a rune path: the keystones, then three rows of minor runes.
 */
export interface RuneRow {
  runes: Array<RuneEntry>;

  /**
   * "keystone", then "row1" to "row3".
   */
  slot: string;
}